Solves _.zip but not without flaws
What: Takes the first array in the list of array arguments and
assumes that length will be the total number of result arrays but it
needs to find the longest array to assume that but I haven't solved
the _.sort function yet.

Why: Fragile without sorting by length of arguments first.